Don't use cleanName
Looks like this function is only used to load the individual songs in an
arrangement.  This results in files not been found as the cleaned up
names don't match the actual names.

It's not very clear if this function is deprecated, but the fact that
it's only used for this particular case, not even for loading regular
songs, makes me think that it is the case.