Dialog boxes interrupt main or option window
The owner window is fetched (usually from the event that triggered the
prompt) and passed through methods call to the dialog box.  I had to
extend many methods to take the owner as argument, because I couldn't
think of a better solution that can be implemented quickly.  Maybe this
could be arranged with a proper controller class, but this is for future
work.