Made safe and unsafe versions of getLine
Some of our issues are somehow tied to the sequence's length---like the
fact that notes can be pasted beyond the displayable limit, or some
freezes occurring because the scrollbar's max value is too low.

The original method manages the IndexOutOfBoundsException by resizing
the sequence automatically.  In some situations, it should be the
responsibility of the class calling getLine to ensure that the
sequence's length is appropriate, and propagate any necessary changes on
the UI side.

For now I just made the unsafe method without using it, but I'm planning
to replace some calls to the safe version in future commits.