Initialize currentLine with -1
Listeners on currentLine will not be triggered at initialization if
currentLine is initialized with 0.  They will also not be triggered for
example when the song is changed but the currentLine is already 0.

Using -1 as a special value for this kind of transitory states solves
many problems by forcing JavaFX to register a change.