New method resetLocation to force redraws
A redraw is triggered whenever a change in currentLine is registered.
This doesn't work in the case where the sequence is changed but the
currentLine is already 0 (its value doesn't change).

The trick to solve this problem is to change currentLine to -1 and then
0.  The change to -1 doesn't trigger a redraw, but the change to 0 does.

The advantage of this over setting up explicit calls to the redraw
method is that there won't be two redraws.  With explicit calls,
there will typically be two redraws (the explicit call and the call from
the listener on currentLine).