Support for additive meters in time signatures
For instance the notation "3+3+2/8" will result in a time signature that
works like "8/8", with the indication that each measure should be
divided in sections of lengths 3, 3 and 2.

The bottom number can still be dropped: "3+3+2".

This changes nothing wrt to how songs are played. Only the length of a
bar is relevant for this.