Staff doesn't hold reference to NoteMatrix
At this point the chain of dependencies between classes that involve
staff display looks like this:
    SMPFXController -> Staff -> StaffImages -> NoteMatrix

Staff has a general redraw method, StaffImages has specialized
initialize and update methods, NoteMatrix manages the notes and
accidentals in particular.

There are some classes elsewhere that need access to NoteMatrix, which
seems wrong to me.  Until the solution is found, I left getters for the
matrix in StaffImages and Staff.