StaffNote doesn't inherit from ImageView
I'm doing this to ensure nodes are only created when they are ready to
be put in the scene.  All functionalities about sequence editing should
managed lighter objects.

The additional call to Staff.redraw is here to make the visual
"selection" effect immediate.

The next commit will do the same thing for accidentals.