Made acc field nonstatic in StaffMouseEventHandler
It appears the reason why this field is static in the first place is to
make the method updateAccidental static and callable from the keyboard
event handler.  This makes it possible to update the silhouette when the
mouse is on the staff and either ctrl of shift is pressed.

But this means updateAccidental actually updates the display, and I'm
trying to move this functionality elsewhere.  So I've changed the method
and made the field nonstatic as well.  This introduces a bug, because
pressing a key while not moving the mouse won't update the silhouette
accidental.  Will fix this in ulterior commits.