Setup resource management with default sf
The new method Utillities.getResourceFile can be used to fetch
resources. A resource is a file that is essential to the program
(sprites, default soundfont, fxml...)

We want users to be able to access resources if they want to customize
the program. For this reason, the method will always try to access the
resource as a regular file; if the file it not found, it is copied from
the program's internal resources.