Replaced ProgramState with SMPMode
This is more manageable. For many situations we only need to know if we
are in song or arranger mode; or if something is playing or not. We
never need to know if a menu is open, since window ownership takes care
of blocking events.

I hope I didn't break something here